requirements.txt
Django==4.0.5
django-cities-light==3.9.1
django-filter==22.1
django-cloudinary-storage==0.3.0
Pillow==9.1.1
dj-database-url==0.5.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
whitenoise==6.2.0
python-decouple==3.6
